requirements-dev.txt
-r requirements.txt
pytest>=7.0
pytest-asyncio>=0.23
pytest-httpx>=0.23
pytest-env>=0.8
black>=22.0
flake8>=5.0
